moving streamLogger function to libp2pUtils.go
